requirements.txt
Babel==2.6.0
Flask==1.0.2
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
GitPython==2.1.11
git+git://github.com/xeBuz/Flask-Validator.git@v1.4#egg=Flask-Validator
pdoc==0.3.2
pytest==4.1.0
pytest-cov==2.7.1
python-dotenv==0.10.1
sadisplay==0.4.9
Werkzeug==0.14.1
stripe==2.17.0
rfc3339==6.0
ansicolors==1.1.8
prometheus_client==0.7.1
